+++ .eslintrc.cjs
@@ -7,7 +7,12 @@ module.exports = {
   },
   parser: '@typescript-eslint/parser',
   parserOptions: {
-    project: './tsconfig.base.json',
+    project: [
+      './tsconfig.base.json',
+      './packages/*/tsconfig.json',
+      './apps/*/tsconfig.json',
+    ],
+    tsconfigRootDir: __dirname,
     sourceType: 'module',
   },
   settings: {